# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Avto Test Imtihon API",
        "description": "Avto Test Imtihon loyihasi uchun API hujjatlari",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Foydalanuvchi to‘lovlarini ro‘yxatini olish",
                "operationId": "afcbc19a07bf33c2fcb73b5c438b476c",
                "responses": {
                    "200": {
                        "description": "To‘lovlar ro‘yxati",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Payment"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Yangi to‘lov yaratish",
                "operationId": "985b848b4cb6ca6115cf4ea71e314c0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "example": "pm_card_visa"
                                    },
                                    "plan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "promocode": {
                                        "type": "string",
                                        "example": "SAVE50",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "To‘lov muvaffaqiyatli yaratildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payment": {
                                            "$ref": "#/components/schemas/Payment"
                                        },
                                        "subscription": {
                                            "$ref": "#/components/schemas/Subscription"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Xato yuz berdi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Promokod yaroqsiz"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Payment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "subscription_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "promocode_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "failed"
                        ],
                        "example": "completed"
                    },
                    "stripe_id": {
                        "type": "string",
                        "example": "ch_test",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-20T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-20T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Subscription": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-20T10:00:00Z"
                    },
                    "ends_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-20T10:00:00Z"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-20T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-20T10:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Payments",
            "description": "Payments"
        }
    ]
}